feat: scroll to top when re-clicking the current route
