+++ src/components/navbar/route/base-route.ts
@@ -2,7 +2,7 @@ import { Badge, Icon, PopupItem } from '@/components/navbar';
 import type { ActionableElement } from '@/lib/event-detection';
 import type { NavbarCard } from '@/navbar-card';
 import type { RouteItemBase } from '@/types';
-import { processTemplate } from '@/utils';
+import { matchesCurrentNavigationPath, processTemplate } from '@/utils';
 
 export class BaseRoute implements ActionableElement {
   private _iconInstance?: Icon;
@@ -60,30 +60,7 @@ export class BaseRoute implements ActionableElement {
           this._navbarCard,
           this.data.selected,
         )
-      : this._browserMatchesURL(this.url);
-  }
-
-  /**
-   * Checks if the current pathname matches the configured URL.
-   * Handles both absolute URLs (starting with "/") and relative URLs (without leading slash).
-   *
-   * @param url - The configured URL (can be absolute or relative)
-   * @returns true if the pathname matches the URL
-   */
-  private _browserMatchesURL(url: string | undefined): boolean {
-    const pathname = window.location.pathname;
-    if (!url) return false;
-
-    if (url.startsWith('/')) {
-      return pathname === url;
-    }
-
-    const normalizedPathname = pathname.endsWith('/')
-      ? pathname.slice(0, -1)
-      : pathname;
-    const normalizedUrl = url.endsWith('/') ? url.slice(0, -1) : url;
-
-    return normalizedPathname.endsWith(`/${normalizedUrl}`);
+      : matchesCurrentNavigationPath(this.url);
   }
 
   get tap_action() {
